Split folder list into separate component
